requirements.txt
Django==2.2.1
django-classic-user-accounts==1.0.32
Pillow==6.0.0
pkg-resources==0.0.0
pytz==2019.1
sorl-thumbnail==12.5.0
sqlparse==0.3.0
django-crispy-forms-1.7.2